feat(iam): determining onchain state for Linea / Verax based only on the value of the score
