+++ app/__tests__/components/NetworkCard.test.tsx
@@ -2,7 +2,6 @@ import React from "react";
 import { screen } from "@testing-library/react";
 
 import { NetworkCard } from "../../components/NetworkCard";
-import { checkOnChainStatus, OnChainStatus } from "../../hooks/useOnChainStatus";
 
 import {
   makeTestCeramicContext,
@@ -11,8 +10,7 @@ import {
 } from "../../__test-fixtures__/contextTestHelpers";
 
 import { UserContextState } from "../../context/userContext";
-import { CeramicContextState, AllProvidersState, ProviderState } from "../../context/ceramicContext";
-import { OnChainProviderType } from "../../context/onChainContext";
+import { CeramicContextState } from "../../context/ceramicContext";
 import { Drawer, DrawerOverlay } from "@chakra-ui/react";
 
 jest.mock("../../utils/onboard.ts");
@@ -53,93 +51,5 @@ describe("OnChainSidebar", () => {
     );
     renderWithContext(mockUserContext, mockCeramicContext, drawer());
     expect(screen.getByText("Sepolia Testnet")).toBeInTheDocument();
-  });
-});
-
-describe("checkOnChainStatus", () => {
-  const issuanceDate0 = new Date();
-  const expirationDate0 = new Date();
-  const issuanceDate1 = new Date();
-  const expirationDate1 = new Date();
-
-  const mockAllProvidersState: AllProvidersState = {
-    ["Google"]: {
-      stamp: {
-        provider: "Google",
-        credential: {
-          expirationDate: expirationDate0,
-          issuanceDate: issuanceDate0,
-          credentialSubject: {
-            hash: "hash1",
-          },
-        },
-      },
-    } as unknown as ProviderState,
-    ["Ens"]: {
-      stamp: {
-        provider: "Ens",
-        credential: {
-          expirationDate: expirationDate1,
-          issuanceDate: issuanceDate1,
-          credentialSubject: {
-            hash: "hash2",
-          },
-        },
-      },
-    } as unknown as ProviderState,
-  };
-
-  const mockOnChainProviders: OnChainProviderType[] = [
-    {
-      providerName: "Google",
-      credentialHash: "hash1",
-      expirationDate: expirationDate0,
-      issuanceDate: issuanceDate0,
-    },
-    {
-      providerName: "Ens",
-      credentialHash: "hash2",
-      expirationDate: expirationDate1,
-      issuanceDate: issuanceDate1,
-    },
-  ];
-
-  const onChainScore = 10.1;
-
-  it("should return NOT_MOVED when onChainProviders is an empty array", () => {
-    expect(checkOnChainStatus(mockAllProvidersState, [], onChainScore, "DONE", onChainScore)).toBe(
-      OnChainStatus.NOT_MOVED
-    );
-  });
-
-  it("should return MOVED_UP_TO_DATE when onChainProviders matches with allProvidersState", () => {
-    expect(checkOnChainStatus(mockAllProvidersState, mockOnChainProviders, onChainScore, "DONE", onChainScore)).toBe(
-      OnChainStatus.MOVED_UP_TO_DATE
-    );
-  });
-
-  it("should return MOVED_OUT_OF_DATE when score does not match", () => {
-    expect(
-      checkOnChainStatus(mockAllProvidersState, mockOnChainProviders, onChainScore + 1, "DONE", onChainScore)
-    ).toBe(OnChainStatus.MOVED_OUT_OF_DATE);
-  });
-
-  it("should return MOVED_OUT_OF_DATE when there are differences between onChainProviders and allProvidersState", () => {
-    const diffMockAllProviderState: AllProvidersState = {
-      ...mockAllProvidersState,
-      ["Github"]: {
-        stamp: {
-          provider: "Github",
-          credential: {
-            credentialSubject: {
-              hash: "hash2",
-            },
-          },
-        },
-      } as unknown as ProviderState,
-    };
-    expect(checkOnChainStatus(diffMockAllProviderState, mockOnChainProviders, onChainScore, "DONE", onChainScore)).toBe(
-      OnChainStatus.MOVED_OUT_OF_DATE
-    );
   });
 });
