+++ app/utils/AttestationProvider.ts
@@ -2,6 +2,13 @@ import onchainInfo from "../../deployments/onchainInfo.json";
 import GitcoinVerifierAbi from "../../deployments/abi/GitcoinVerifier.json";
 import axios, { AxiosResponse } from "axios";
 import { iamUrl } from "../config/stamp_config";
+import { OnChainStatus } from "../hooks/useOnChainStatus";
+import { ScoreStateType } from "../context/scorerContext";
+import { AllProvidersState, ProviderState } from "../context/ceramicContext";
+import { OnChainProviderType } from "../context/onChainContext";
+import { Stamp } from "@gitcoin/passport-types";
+
+type ProviderWithStamp = ProviderState & { stamp: Stamp };
 
 type AttestationProviderStatus = "enabled" | "comingSoon" | "disabled";
 
@@ -30,6 +37,13 @@ export interface AttestationProvider {
   verifierAddress: () => string;
   verifierAbi: () => any;
   getMultiAttestationRequest: (payload: {}) => Promise<AxiosResponse<any, any>>;
+  checkOnChainStatus: (
+    allProvidersState: AllProvidersState,
+    onChainProviders: OnChainProviderType[],
+    rawScore: number,
+    scoreState: ScoreStateType,
+    onChainScore: number
+  ) => OnChainStatus;
 }
 
 class BaseAttestationProvider implements AttestationProvider {
@@ -70,6 +84,47 @@ class BaseAttestationProvider implements AttestationProvider {
       transformRequest: [(data: any) => JSON.stringify(data, (_k, v) => (typeof v === "bigint" ? v.toString() : v))],
     });
   }
+
+  checkOnChainStatus(
+    allProvidersState: AllProvidersState,
+    onChainProviders: OnChainProviderType[],
+    rawScore: number,
+    scoreState: ScoreStateType,
+    onChainScore: number
+  ): OnChainStatus {
+    // This is default implementation that will check for differences in
+    // the on-chain providers and on-chain score
+    if (scoreState !== "DONE") return OnChainStatus.LOADING;
+
+    if (onChainProviders.length === 0) return OnChainStatus.NOT_MOVED;
+
+    if (rawScore !== onChainScore) return OnChainStatus.MOVED_OUT_OF_DATE;
+
+    const verifiedDbProviders: ProviderWithStamp[] = Object.values(allProvidersState).filter(
+      (provider): provider is ProviderWithStamp => provider.stamp !== undefined
+    );
+
+    const [equivalentProviders, differentProviders] = verifiedDbProviders.reduce(
+      ([eq, diff], provider): [ProviderWithStamp[], ProviderWithStamp[]] => {
+        const expirationDateSeconds = Math.floor(new Date(provider.stamp.credential.expirationDate).valueOf() / 1000);
+        const issuanceDateSeconds = Math.floor(new Date(provider.stamp.credential.issuanceDate).valueOf() / 1000);
+
+        const isEquivalent = onChainProviders.some(
+          (onChainProvider) =>
+            onChainProvider.providerName === provider.stamp.provider &&
+            onChainProvider.credentialHash === provider.stamp.credential.credentialSubject?.hash &&
+            Math.floor(onChainProvider.expirationDate.valueOf() / 1000) === expirationDateSeconds &&
+            Math.floor(onChainProvider.issuanceDate.valueOf() / 1000) === issuanceDateSeconds
+        );
+        return isEquivalent ? [[...eq, provider], diff] : [eq, [...diff, provider]];
+      },
+      [[], []] as [ProviderWithStamp[], ProviderWithStamp[]]
+    );
+
+    return equivalentProviders.length === onChainProviders.length && differentProviders.length === 0
+      ? OnChainStatus.MOVED_UP_TO_DATE
+      : OnChainStatus.MOVED_OUT_OF_DATE;
+  }
 }
 
 export class EASAttestationProvider extends BaseAttestationProvider {
@@ -105,5 +160,20 @@ export class VeraxAndEASAttestationProvider extends EASAttestationProvider {
       },
       transformRequest: [(data: any) => JSON.stringify(data, (_k, v) => (typeof v === "bigint" ? v.toString() : v))],
     });
+  }
+
+  checkOnChainStatus(
+    allProvidersState: AllProvidersState,
+    onChainProviders: OnChainProviderType[],
+    rawScore: number,
+    scoreState: ScoreStateType,
+    onChainScore: number
+  ): OnChainStatus {
+    // This is specific implementation for Verax where we only check for the score to be different
+    if (scoreState !== "DONE" || onChainScore === undefined) {
+      return OnChainStatus.LOADING;
+    }
+    if (Number.isNaN(onChainScore)) return OnChainStatus.NOT_MOVED;
+    return rawScore !== onChainScore ? OnChainStatus.MOVED_OUT_OF_DATE : OnChainStatus.MOVED_UP_TO_DATE;
   }
 }
